feat: evidence of learning for tutorial 11
set up basic files for tutorial 11.

add basic HTML and CSS for tutorial 11.

add <blockquote> element and apply a float property to cause the content to wrap around the floating content.

style  flexible width, resizable images, and media queries to cause the image to float to the right at a specific screen width, and then at a larger screen size pull the image into the right side margin.

add columns with gap and divider and remove breaks in column content for better flow.

use media queries, columns, and negative margins to create an ordered list that organizes content in the column

add summary of learning for styling content with float and columns.

FSA22V2-60