feat: evidence of learning for tutorial 19
add index.html, styles.css, README.md, and base HTML and CSS code  for tutorial 19.

add an image folder with image in it, add image tag to html, give it an alt tag, and set the max-width 100% on the img element so that it scales with browser window size.

add an image caption with figcaption element, position it absolute to the relatively positioned parent figure element, set it to inline-block so element will shrink to fit its contents and set images that are in figure elements to display: block.

add an image caption with figcaption element, position it absolute to the relatively positioned parent figure element, set it to inline-block so element will shrink to fit its contents and set images that are in figure elements to display: block.

Also use responsive image swapping by uploading a series of images of different sizes, linking them as source elements in a picture element, and setting min-width in media property to an appropriate min-width for the image. Set object-fit in .hero-image class to cover and also object-position to bottom-right.

Add summary of learning for  “How To Style HTML Elements with Borders, Shadows, and Outlines in CSS” Tutorial 19.

FSA22V2-65