feat: evidence of learning for tutorial 18
Add base HTML, CSS, and README.md files for tutorial 18.

add in css: modal styling, set opacity to zero on .modal-container, set opacity to 1 on
.modal-container:target, in index.html create id of “disclaimer” on .modal-container class
and apply the id to href link on button, in css: set pointer-events to “none” on
.modal-container to make it invisible by default and pointer-events to “all” to make it
visible when targeted, and cause modal to fade in and out of view by adding transition
property to the .modal-container.

distinguish modal header, body, and border with varying alpha channel transparency
values for background color,  use alpha channel values to make drop-shadow on .modal
more subtle by using slightly more transparency on the box-shadow property.

hide “close” link in .modal-close with transparent color value, add an X shape to replace
close link on modal using overlapping 45 degree-angle linear gradients, create hover and
focus states for .modal-close to better cue user when interacting with the X to close the
modal.

add background-image with different colors from the button:hover and button:focus, then
add transition on background-color with time and ease to cause gradient to switch on
hover.

Add summary of learning for  “How To Use Opacity and Transparency to Create a Modal in
CSS” Tutorial 18.

FSA22V2-65